+++ CHANGES.md
@@ -1,3 +1,9 @@
+# UNRELEASED
+
+
+- Builtins:
+  - Add `choose`, `min`, `max` and `partition` methods to OCaml sets
+
 # v1.19.1 (June 2024)
 
 Requires Menhir 20211230 and OCaml 4.08 or above.